+++ cli/commands/event/show.go
@@ -89,7 +89,7 @@ func printEntityToList(event *types.Event, writer io.Writer) {
 			},
 			{
 				Label: "Silenced",
-				Value: strconv.FormatBool(len(event.Silenced) > 0),
+				Value: strconv.FormatBool(len(event.Check.Silenced) > 0),
 			},
 			{
 				Label: "Timestamp",
@@ -98,10 +98,10 @@ func printEntityToList(event *types.Event, writer io.Writer) {
 		},
 	}
 
-	if len(event.Silenced) > 0 {
+	if len(event.Check.Silenced) > 0 {
 		silencedBy := &list.Row{
 			Label: "Silenced By",
-			Value: strings.Join(event.Silenced, ", "),
+			Value: strings.Join(event.Check.Silenced, ", "),
 		}
 		cfg.Rows = append(cfg.Rows[:len(cfg.Rows)-1], silencedBy, cfg.Rows[len(cfg.Rows)-1])
 	}