+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Changed the maximum number of open file descriptors on a system to from 1024
 (default) to 65535.
 - Increased the default etcd size limit from 2GB to 4GB.
+- Move Hooks and Silenced out of Event and into Check.
+
 ### Added
 - Support for managing mutators via sensuctl.